OBSIDIAN EMPRESA/app/backend/requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
pgvector==0.2.4
python-dotenv==1.0.1
pydantic==2.6.0
openai==1.12.0
anthropic==0.19.1
python-multipart==0.0.9
